requirements.txt
streamlit
requests
plotly
google-generativeai